+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.49"
+__version__ = "0.12.50"
 
 
 def safe_commit() -> None: